SMARS Line_Following_Robot
A smart robot line follower using Arduino and three IR sensors is designed to automatically follow a predefined path on the ground. The IR sensors detect the contrast between the line and the surface, allowing the robot to identify whether it is on or off the track. The Arduino processes the sensor data and controls the motors to adjust the robot’s direction in real time. With three sensors, the robot can accurately detect curves and make smoother turns. This project demonstrates the basic principles of automation, sensing, and embedded system control.